+++ cardano-client-demo/LedgerState.hs
@@ -33,7 +33,7 @@ main = do
   -- Get socket path from CLI argument.
   configFilePath : socketPath : xs <- getArgs
   blockCount <- fmap (either (error . T.unpack . renderFoldBlocksError) id) $ runExceptT $ foldBlocks
-    configFilePath
+    (File configFilePath)
     (File socketPath)
     FullValidation
     (0 :: Int) -- We just use a count of the blocks as the current state