+++ cardano-testnet/test/Test/FoldBlocks.hs
@@ -17,11 +17,12 @@ import qualified Hedgehog.Extras.Stock.IO.Network.Sprocket as HE
 import qualified Hedgehog.Extras.Test as HE
 import qualified Hedgehog.Extras.Test.Base as H
 
+import           Cardano.Api (File (..))
 import qualified Cardano.Api as C
 import           Cardano.Testnet as TN
+
 import qualified Testnet.Util.Base as H
 import           Testnet.Util.Runtime
-
 
 newtype FoldBlocksException = FoldBlocksException C.FoldBlocksError
 instance Exception FoldBlocksException
@@ -69,7 +70,7 @@ prop_foldBlocks = H.integrationRetryWorkspace 2 "foldblocks" $ \tempAbsBasePath'
       -- that case we simply restart `foldBlocks` again.
       forever $ do
         let handler _env _ledgerState _ledgerEvents _blockInCardanoMode _ = IO.putMVar lock ()
-        e <- runExceptT (C.foldBlocks configFile (C.File socketPathAbs) C.QuickValidation () handler)
+        e <- runExceptT (C.foldBlocks (File configFile) (C.File socketPathAbs) C.QuickValidation () handler)
         either (throw . FoldBlocksException) (\_ -> pure ()) e
     link a -- Throw async thread's exceptions in main thread